requirements.txt
Django==1.4
South==0.7.5
django-bootstrap-form==2.0.3
django-registration==0.8
django-social-auth==0.6.9
django-tastypie==0.9.11
httplib2==0.7.4
mimeparse==0.1.3
oauth2==1.5.211
python-dateutil==1.5
python-openid==2.2.5
wsgiref==0.1.2
django-autoslug==1.5.0
-e git+https://github.com/spreeker/django-voting2.git#egg=django-voting